tests: merge `server_test.py` to `test_utils.py`
